requirements.txt
torch
torchvision
Pillow
numpy
clip @ git+https://github.com/openai/CLIP.git
PyQt5